Move AI card (right side) to another component
